requirements.txt
requests==2.28.1
urllib3==1.26.9
python-dotenv==0.20.0